- Updates to use new Prescat project table from late July - adds 3 fields, renames Proj_zip field (changed case from Proj_Zip), updates manifest
- Updates all files in manifest to use new S3 bucket on the codefor dc account: `housing-insights` (instead of without the dash, which was bucket name on Neal's personal S3 account)
- Fixes project loading bug re: incomplete refactor of fields needed for rename-census-tracts
- Fixes bugs related to new ProjectBaseApiConn - split into separate file to avoid more circular import issues.
