+++ python/housinginsights/sources/models/DCHousing.py
@@ -19,30 +19,56 @@ SUBSIDY_FIELDS_MAP = {'Nlihc_id': None, 'Subsidy_id': None,
                       'Subsidy_info_source_property': None,
                       'POA_end_actual': None}
 
-PROJECT_FIELDS_MAP = {'Nlihc_id': None, 'Status': None, 'Subsidized': None,
-                      'Cat_Expiring': None, 'Cat_Failing_Insp': None,
-                      'Proj_Name': 'PROJECT_NAME', 'Proj_City': None,
-                      'Proj_ST': None, 'Proj_Zip': None, 'Proj_Units_Tot': None,
+PROJECT_FIELDS_MAP = {'Nlihc_id': None,
+                      'Status': None, 
+                      'Subsidized': None,
+                      'Cat_Expiring': None, 
+                      'Cat_Failing_Insp': None,
+                      'Proj_Name': 'PROJECT_NAME', 
+                      'Proj_City': None,
+                      'Proj_ST': None, 
+                      'Proj_zip': None, 
+                      'Proj_Units_Tot': None,
                       'Proj_Units_Assist_Min': "REPORT_UNITS_AFFORDABLE",
                       'Proj_Units_Assist_Max': "REPORT_UNITS_AFFORDABLE",
-                      'Hud_Own_Effect_dt': None, 'Hud_Own_Name': None,
-                      'Hud_Own_Type': None, 'Hud_Mgr_Name': None,
-                      'Hud_Mgr_Type': None, 'Subsidy_Start_First': None,
-                      'Subsidy_Start_Last': None, 'Subsidy_End_First': None,
-                      'Subsidy_End_Last': None, 'Ward2012': 'MAR_WARD',
-                      'PBCA': None, 'Anc2012': None, 'Psa2012': None,
-                      'Geo2010': None, 'Cluster_tr2000': None,
-                      'Cluster_tr2000_name': None, 'Zip': None,
-                      'Proj_image_url': None, 'Proj_streetview_url': None,
-                      'Proj_address_id': 'ADDRESS_ID', 'Proj_x': 'XCOORD',
-                      'Proj_y': 'YCOORD', 'Proj_lat': 'LATITUDE',
+                      'Hud_Own_Effect_dt': None, 
+                      'Hud_Own_Name': None,
+                      'Hud_Own_Type': None, 
+                      'Hud_Mgr_Name': None,
+                      'Hud_Mgr_Type': None, 
+                      'Subsidy_Start_First': None,
+                      'Subsidy_Start_Last': None, 
+                      'Subsidy_End_First': None,
+                      'Subsidy_End_Last': None, 
+                      'Ward2012': 'MAR_WARD',
+                      'PBCA': None, 
+                      'Anc2012': None, 
+                      'Psa2012': None,
+                      'Geo2010': None, 
+                      'Cluster_tr2000': None,
+                      'Cluster_tr2000_name': None, 
+                      'Zip': None,
+                      'Proj_image_url': None, 
+                      'Proj_streetview_url': None,
+                      'Proj_address_id': 'ADDRESS_ID', 
+                      'Proj_x': 'XCOORD',
+                      'Proj_y': 'YCOORD', 
+                      'Proj_lat': 'LATITUDE',
                       'Proj_lon': 'LONGITUDE',
                       'Update_Dtm': 'GIS_LAST_MOD_DTTM',
                       'Subsidy_info_source_property': None,
-                      'contract_number': None, 'Proj_addre': 'FULLADDRESS',
-                      'Bldg_count': None, 'Category_Code': None,
-                      'Cat_At_Risk': None, 'Cat_More_Info': None,
-                      'Cat_Lost': None, 'Cat_Replaced': None}
+                      'contract_number': None, 
+                      'Proj_addre': 'FULLADDRESS',
+                      'Proj_ayb': None,
+                      'Proj_eyb': None,
+                      'Proj_owner_type':None,
+                      'Bldg_count': None, 
+                      'Category_Code': None,
+                      'Cat_At_Risk': None, 
+                      'Cat_More_Info': None,
+                      'Cat_Lost': None, 
+                      'Cat_Replaced': None
+                      }
 
 
 class DCHousingResult(object):
