+++ python/housinginsights/sources/models/dhcd.py
@@ -101,7 +101,7 @@ PROJECT_FIELDS_MAP = {  'Nlihc_id': None,
                         'Proj_Name': 'property_name', 
                         'Proj_City': None,
                         'Proj_ST': None, 
-                        'Proj_Zip': None, 
+                        'Proj_zip': None, 
                         'Proj_Units_Tot': None,
                         'Proj_Units_Assist_Min': "units__total_number_of_affordable_units",
                         'Proj_Units_Assist_Max': "units__total_number_of_affordable_units",
@@ -133,6 +133,9 @@ PROJECT_FIELDS_MAP = {  'Nlihc_id': None,
                         'Subsidy_info_source_property': None,
                         'contract_number': None, 
                         'Proj_addre': 'address_single', #calculated during create_project_subsidy_csv
+                        'Proj_ayb': None,
+                        'Proj_eyb': None,
+                        'Proj_owner_type':None,
                         'Bldg_count': None, 
                         'Category_Code': None,
                         'Cat_At_Risk': None,
