+++ python/housinginsights/sources/models/pres_cat.py
@@ -3,7 +3,7 @@ Model for tables from preservation catalog.
 """
 
 PROJ_FIELDS = ['Nlihc_id', 'Status', 'Subsidized', 'Cat_Expiring',
-          'Cat_Failing_Insp', 'Proj_Name', 'Proj_City', 'Proj_ST', 'Proj_Zip',
+          'Cat_Failing_Insp', 'Proj_Name', 'Proj_City', 'Proj_ST', 'Proj_zip',
           'Proj_Units_Tot', 'Proj_Units_Assist_Min', 'Proj_Units_Assist_Max',
           'Hud_Own_Effect_dt', 'Hud_Own_Name', 'Hud_Own_Type', 'Hud_Mgr_Name',
           'Hud_Mgr_Type', 'Subsidy_Start_First', 'Subsidy_Start_Last',
@@ -12,14 +12,14 @@ PROJ_FIELDS = ['Nlihc_id', 'Status', 'Subsidized', 'Cat_Expiring',
           'Cluster_tr2000_name', 'Zip', 'Proj_image_url', 'Proj_streetview_url',
           'Proj_address_id', 'Proj_x', 'Proj_y', 'Proj_lat', 'Proj_lon',
           'Update_Dtm', 'Subsidy_info_source_property', 'contract_number',
-          'Proj_addre', 'Bldg_count', 'Category_Code',
+          'Proj_addre', 'Bldg_count','Proj_ayb', 'Proj_eyb', 'Proj_owner_type', 'Category_Code',
           'Cat_At_Risk', 'Cat_More_Info', 'Cat_Lost', 'Cat_Replaced']
 
 ADDRESS_FIELDS = {'lat_lon': ('Proj_lat', 'Proj_lon'),
                   'xy_coords': ('Proj_x', 'Proj_y'),
-                  'address': ('Proj_addre', 'Proj_Zip')}
+                  'address': ('Proj_addre', 'Proj_zip')}
 
-MAR_MAP = {'Proj_Name', 'Proj_City', 'Proj_ST', 'Proj_Zip',
+MAR_MAP = {'Proj_Name', 'Proj_City', 'Proj_ST', 'Proj_zip',
           'Ward2012', 'PBCA',
           'Anc2012', 'Psa2012', 'Cluster_tr2000',
           'Cluster_tr2000_name', 'Zip', 'Proj_image_url', 'Proj_streetview_url',
